refactor(auth): extract callback handler to lib to avoid route file testing
Move handleCallback, buildSessionCookie, CLEAR_STATE_COOKIE, and the GET
handler wiring into src/lib/auth/callback-handler.ts so tests live in the
lib tree instead of the routes tree. This eliminates the TanStack Router
warning about a route file that doesn't export a Route.
